+++ src/graph/access.rs
@@ -331,7 +331,7 @@ impl<'a> Analyzer<'a> {
     pub fn reads(&self, source: &Source, deps: &Deps) -> Reads {
         let mut out = self.roots.reads(source);
         for path in deps.files() {
-            if path.extension().is_some_and(|ext| ext == Config::TYPST) {
+            if Config::has_ext(path, Config::TYPST) {
                 out.extend(self.file(path).iter().cloned());
             }
         }